test(study): validates resource group creation
