+++ src/Sepes.RestApi/Startup.cs
@@ -185,7 +185,7 @@ namespace Sepes.RestApi
             services.AddTransient<IAzureResourceGroupService, AzureResourceGroupService>();
             services.AddTransient<IAzureNetworkSecurityGroupService, AzureNetworkSecurityGroupService>();
             services.AddTransient<IAzureBastionService, AzureBastionService>();
-            services.AddTransient<IAzureVirtualNetworkService, AzureVNetService>();
+            services.AddTransient<IAzureVirtualNetworkService, AzureVirtualNetworkService>();
             services.AddTransient<IAzureVirtualMachineService, AzureVirtualMachineService>();
             services.AddTransient<IAzureVirtualMachineExtenedInfoService, AzureVirtualMachineExtendedInfoService>();
             services.AddTransient<IAzureQueueService, AzureQueueService>();
@@ -196,7 +196,7 @@ namespace Sepes.RestApi
             services.AddTransient<IAzureNetworkSecurityGroupRuleService, AzureNetworkSecurityGroupRuleService>();
             services.AddTransient<IAzureResourceSkuService, AzureResourceSkuService>();
             services.AddTransient<IAzureUserService, AzureUserService>();
-            services.AddTransient<IAzureVmOsService, AzureVmOsService>();
+            services.AddTransient<IAzureVirtualNetworkOperatingSystemService, AzureVirtualNetworkOperatingSystemService>();
             services.AddTransient<IAzureCostManagementService, AzureCostManagementService>();
             services.AddTransient<IAzureRoleAssignmentService, AzureRoleAssignmentService>(); 
         }
