+++ src/Sepes.Tests/Setup/BasicServiceCollectionFactory.cs
@@ -58,7 +58,7 @@ namespace Sepes.Tests.Setup
             services.AddTransient<IAzureResourceGroupService, AzureResourceGroupService>();
             services.AddTransient<IAzureNetworkSecurityGroupService, AzureNetworkSecurityGroupService>();
             services.AddTransient<IAzureBastionService, AzureBastionService>();
-            services.AddTransient<IAzureVirtualNetworkService, AzureVNetService>();
+            services.AddTransient<IAzureVirtualNetworkService, AzureVirtualNetworkService>();
             services.AddTransient<IAzureVirtualMachineService, AzureVirtualMachineService>();
             services.AddTransient<IAzureQueueService, AzureQueueServiceMock>();
             services.AddTransient<IAzureStorageAccountService, AzureStorageAccountService>();
